ai_server/requirements.txt
# Core AI and Computer Vision
opencv-python==4.8.1.78
dlib==19.24.2
numpy==1.24.3
tensorflow==2.13.0
tensorflow-lite==2.13.0

# Web Framework and API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0

# Image Processing
Pillow==10.0.1
imutils==0.5.4

# Database and Caching
pymongo==4.5.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
asyncio==3.4.3
aiofiles==23.2.1

# Performance and Monitoring
psutil==5.9.6
prometheus-client==0.19.0

# Google Cloud Storage
google-cloud-storage==2.10.0

# JWT Authentication
PyJWT==2.8.0

# HTTP Client
httpx==0.25.2

# Data Processing
pandas==2.1.4
scikit-learn==1.3.2

